+++ config.js
@@ -4,6 +4,6 @@ module.exports = {
   testFiles: [
     '**/__testUtils__/**/*.{js,jsx,ts,tsx}',
     '**/__tests__/**/*.{js,jsx,ts,tsx}',
-    '**/*.{spec,test}.{js,jsx,ts,tsx}'
+    '**/*.{spec,test,e2e,cy}.{js,jsx,ts,tsx}'
   ]
 };