+++ react.js
@@ -44,6 +44,13 @@ module.exports = {
       ERROR,
       {props: 'never', children: 'ignore'}
     ],
+    'react/function-component-definition': [
+      ERROR,
+      {
+        namedComponents: 'function-declaration',
+        unnamedComponents: 'arrow-function'
+      }
+    ],
     'react/jsx-handler-names': [
       ERROR,
       {
@@ -56,6 +63,7 @@ module.exports = {
       {allowArrowFunctions: true, allowFunctions: true}
     ],
     'react/jsx-no-target-blank': ERROR,
+    'react/jsx-no-useless-fragment': ERROR,
     'react/jsx-sort-props': [
       ERROR,
       {
@@ -67,6 +75,7 @@ module.exports = {
     'react/no-direct-mutation-state': ERROR,
     'react/no-this-in-sfc': ERROR,
     'react/no-typos': ERROR,
+    'react/no-unstable-nested-components': [ERROR, {allowAsProps: true}],
     'react/no-unused-prop-types': ERROR,
     'react/no-unused-state': ERROR,
     'react/self-closing-comp': ERROR,